pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Taxi demand predictor service"
authors = ["Marc-Steffen Kaesz <mkaesz@msk.pub>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
jupyter = "^1.0.0"
requests = "^2.31.0"
pyarrow = "^14.0.0"
tqdm = "^4.66.1"
plotly = "^5.18.0"
scikit-learn = "^1.3.2"
xgboost = "^2.0.1"
sklearn = "^0.0.post11"
lightgbm = "^4.1.0"
optuna = "^3.4.0"
python-dotenv = "^1.0.0"
setuptools = "^68.2.2"
hopsworks = "^3.4.3"
streamlit = "^1.28.1"
geopandas = "^0.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
